requirements.txt
absl-py==1.0.0
appnope==0.1.2
arrow==1.2.2
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
black==22.1.0
blessed==1.19.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.0.0
drf-excel==1.0.0
drf-renderer-xlsx==1.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
executing==0.8.2
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.45.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
peewee==3.14.10
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
swagger-spec-validator==2.7.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.8
waitress==2.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.7.0
